:ok_hand: [#15] Make upper bound of datetime filters exclusive
The wording in the sample API specification from Felix clearly
indicates that the upper bound should not be inclusive (lt vs lte).

For creatiedatum, nothing was specified, but it's a bit of an odd
one because it's a date field rather than datetime, which is
more ambiguous. Here it's more natural to include the upper
bound, and this has been made clear with the parameter name
'tot en met'.

VCR cassettes are re-recorded and code updated, and the API
spec is updated too.