:sparkles: [#16] Filter and aggregate on document information categories
To ensure that the name/uuid pairs stay together, they must be
defined as nested field in the mappings, but that nested nature
makes the filter and aggregations a bit more convoluted than you'd
ideally like.

We must aggregate on the nested object and within that run another
aggregation to get the bucket of uuid/name keys.