:sparkles: [#27] Load self-signed certificates/private CA bundles
The setup function is tweaked so that it only runs once, as importing the
celery tasks triggers an import of the celery app, which itself also
calls setup_env (required for the Celery worker entrypoint).

The setup function loads the self-signed certificates and updates the
environment. This is done with multi-threading in mind (when running
with multiple threads in uwsgi/containers), so locks are used to
prevent race conditions when updating the environment.

The environment is global for the whole process and shared by all
threads, so as long as one thread succeeds, this should be sufficient.