:sparkles: [#27] Support custom CA in ES client
Grab the (either) explicitly configured CA for ES only, or fall back
to EXTRA_VERIFY_CERTS if loaded via self-certifi for a 'blanket'
solution.

This isn't really easy to unit test unless we do a whole custom
CA setup purely to check the connection verification, so we leave
it to the infra side to verify that it works correctly.