:test_tube: Add failing test for desired aggregation/buckets responses
When a filter is applied, we want to keep seeing the buckets for the
unfiltered dataset (single filter). However, when a filter for another
bucket is used, that should affect the returned buckets for the other
filters.

If there are three publishers A, B and C, and three information
categories X, Y and Z, then:

* when filtering on publisher B, the amount of available categories
  must be restricted to results matching publisher B
* when filtering on category Z, then the available publishers should
  be restricted to results having category Z, e.g. publishers A and B

In a future commit, the result type will also be incorporated, but that
one is a bit different because we need to take the _index meta field
into account rather than limiting the search to a particular index.