:bug: [#61] Fix boolean operators
The query_string search in ES supports AND/OR behaviour naturally,
but this does not apply for simple_query_string which only
understands + and |. So, we must convert the syntax before
we feed it to ES.

This uses a regex approach aimed at not destroying quoted
phrases. The input is taken as-is, respecting any
parentheses used for precedence and processes each bit of
the query string, only replacing AND/OR words with their
operator equivalents.

ES documentation gives examples where the term directly
follows the operator, e.g. +foo, but it appears that
spaces in between (like + foo) work fine too.