:boom: [#62] Switch resultType field to resultTypes array field
Anticipating the future where an additional document type will be
introduced, and in light of the bucket aggregation changes made
earlier, it was agreed to make the field an array of strings.

If no filter is provided (or empty array), then no filtering is
applied, otherwise results are restricted to the specified document
types.