:art: [#81] Clean up implementation
Cleaned up the implementation of the zip/7z archive processing and
fixed a file size bug.

file.__sizeof__() returns the memory usage of an object, not the size
of the binary content contained inside. For that, use the length of
the bytes obtained via .read. We have to call .read anyway to base64
encode it, so there's no performance loss there.

Furthermore, the code duplication to get the result has been inverted
by using generators - all we really want from the archives processing
is that they give use file-like objects for us to inspect and process,
so a generator approach to yield files one-by-one allows for a single
function that processes files while composing the iterators that
provide the files to process. This way, we don't have to pass around
calculated total file sizes and can keep the core logic centralized.