:sparkles: [#31] added kenmerken to the document/publication
Added support to index the kenmerken from the gpp-zoeken.
We only index the the kenmerk value of it and not the bron,
because this isn't relevant for civilians. The index will
be in the form of a list.

Also added the function that the kenmerken field if empty
will be filled in by the identifier (only on document)

And changed the identifier field in the query search for identifiers.