:white_check_mark: [#100] fix wrongly asserted test
The test is set to relevance, which prioritizes the '_score' over the
'laatst_gewijzigd_datum'.
The '_score' in both cases is 0.0 so the determinine factor is the 'laatst_gewijzigd_datum' field which is set to descending order.
The test initially asserted that 2025-01-15 should come before 2025-01-26 which would be ascending order